Add files via upload
This is a run, walk or bike calculator with 3 Inputs: Distance, Time & Pace. If you know 2 of them and post those the 3rd one will be the answer from the other two. You can use before an event to tell you what pace you need to maintain to get the time you want or did, etc.